test(wan2_2): cast latents to bf16 at VAE decoder boundary
Same pattern as the DiT boundary in _denoise — VAE decoder has bf16
weights, so float32 latents (from the new float32 denoise loop) must
be cast at the call site. Unscaling still happens in float32 for
precision.

Caught in verification run: RuntimeError: Input type (float) and bias
type (c10::BFloat16) should be the same at post_quant_conv.

+++ tests/torch/models/wan2_2/test_wan22_e2e.py
@@ -240,10 +240,12 @@ def _decode_and_save(vae, latents: torch.Tensor, out_path: Path) -> None:
     ).to(latents.dtype)
     latents_unscaled = latents / latents_std + latents_mean
 
+    # VAE decoder has bf16 weights — cast at the boundary (same pattern as
+    # the DiT call in _denoise). Unscaling happens in float32 for precision.
     decoder_wrapper = VAEDecoderWrapper(vae).eval().bfloat16()
     pixels = run_component(
         decoder_wrapper,
-        [latents_unscaled],
+        [latents_unscaled.to(torch.bfloat16)],
         on_tt=TT_VAE_DECODER,
         shard_spec_fn=(lambda m: shard_vae_decoder_specs(m.vae)),
     )
